chore: add react kitchen sink navigation
